requirements .txt
anyio==4.2.0
APScheduler==3.6.3
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.4
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
lxml==5.1.0
pillow==10.2.0
python-telegram-bot==13.7
pytz==2023.3.post1
requests==2.31.0
schedule==1.2.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tornado==6.4
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
zope.interface==6.1
PyYAML==6.0